refactor: pks/testutils/fake
